pyproject.toml
    [project]
    name = "todo-list"
    version = "0.1.0"
    description = ""
    packages = [{include = "src"}]
    authors = [
        {name = "nitin soni",email = "nitinsoni815@gmail.com"}
    ]
    readme = "README.md"
    requires-python = ">=3.12"
    dependencies = [
        "flask (>=3.1.0,<4.0.0)",
        "flask-restful (>=0.3.10,<0.4.0)",
        "flask-sqlalchemy (>=3.1.1,<4.0.0)",
        "sqlalchemy (>=2.0.38,<3.0.0)",
        "python-dotenv (>=1.0.1,<2.0.0)",
        "psycopg2 (>=2.9.10,<3.0.0)"
    ]


    [build-system]
    requires = ["poetry-core>=2.0.0,<3.0.0"]
    build-backend = "poetry.core.masonry.api"
    

[tool.poetry]
package-mode = false